+++ changelog.md
@@ -8,6 +8,7 @@
    - Added support to future time (tomorrow / in X minutes or seconds)
    - Added tooltip on hover showing full datetime
 3. Added operation and user search in audit log
+4. Added compare discord members
 
 ## v3.4.3